Get rest of make_segments and crash_and_concern working
