+++ data_generation/README.md
@@ -23,12 +23,14 @@ Although it's not necessary, QGIS (http://www.qgis.org/en/site/forusers/download
 
 ## Process map
 
+All of the data generation scripts should be run from the top directory (boston-crash-modeling) using the following scheme: `python -m <import path> <args>`.  See below for specific examples.
+
 ## 1) Extract intersections
 - Reads in road segment data (data/raw/Boston_Segments.shp)
 - Finds point locations where roads intersect
 - Creates a shapefile of intersections (inters.shp)
-- <b>Usage:</b> python -m make_segments.extract_intersections ../data/raw/Boston_Segments.shp
-- <b>Results:</b>
+- **Usage:** python -m data_generation.make_segments.extract_intersections data/raw/Boston_Segments.shp
+- **Results:**
     - data/processed/maps/inters.shp (and related files)
 
 
@@ -39,12 +41,12 @@ Although it's not necessary, QGIS (http://www.qgis.org/en/site/forusers/download
     - Road features from connected road segments linked to intersection id (for later aggregation)
 - Separates out non-intersection segments
 -Creates unique segment ids where all non-intersections have a '00' prefix <br>
-- <b>Usage:</b> python create_segments.py
-- <b>Dependencies:</b>
+- **Usage:** python -m data_generation.make_segments.create_segments
+- **Dependencies:**
     - data/processed/maps/inters.shp
     - data/processed/maps/ma\_co\_spatially\_joined\_streets.shp
         - Descriptions of the attributes from ma_co_spatially_joined_streets.shp can be found in data/docs/MassDOTRoadInvDictionary.pdf
-- <b>Results:</b>
+- **Results:**
     - data/processed/inters_segments.shp
     - data/processed/non_inters_segments.shp
     - data/processed/inters_data.json
@@ -56,7 +58,7 @@ Although it's not necessary, QGIS (http://www.qgis.org/en/site/forusers/download
     - Tolerance of 30m for crashes, 20m for concerns
 - Writes shapefile of joined points and json data file
 - Includes coordinates and near_id referring to segment id (intersection or non intersection)
-- <b>Usage:</b> python join_segments_crash_concern.py
+- <b>Usage:</b> python -m data_generation.crash_and_concern.join_segments_crash_concern
 - <b>Dependencies:</b>
     - inters/non_inters shape data
     - CAD crash data: data/raw/cad_crash_events_with_transport_2016_wgs84.csv
