+++ data_generation/crash_and_concern/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package joins crash and concern data to road segments
+"""
