+++ data_generation/make_segments/create_segments.py
@@ -14,7 +14,7 @@ from fiona.crs import from_epsg
 from shapely.geometry import Point, shape, mapping
 from shapely.ops import unary_union
 from collections import defaultdict
-from make_segments_util import write_shp
+from ..util import write_shp
 
 
 def get_intersection_buffers(intersections, intersection_buffer_units):
@@ -53,8 +53,8 @@ def find_non_ints(roads):
 
 if __name__ == '__main__':
 
-    MAP_FP = '../../data/processed/maps'
-    DATA_FP = '../../data/processed'
+    MAP_FP = 'data/processed/maps'
+    DATA_FP = 'data/processed'
 
     print "Map data at ", MAP_FP
     print "Output intersection data to ", DATA_FP
